cli: Use host prefixes as unique identifiers
Using the contact ID (which is about to disappear, anyway) to refer to
contacts in commands makes for an awkward UX.

For a differently awkward UX, use the shortest (>3) unambiguous prefix
of the hostname instead. This is stable, always typable, kind of
memorable, and seems like a decent compromise for now.